requirements.txt
fastapi==0.110.0
uvicorn==0.29.0
pydantic==2.6.4
dnspython==2.6.1
email-validator==2.2.0
aiofiles==23.2.1
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
validate_email>=1.3.0
pandas>=1.5.0
python-dotenv>=1.0.0
python-multipart
